Renaming iree/base/signature_mangle.* to signature_parser.*.
