+++ bindings/python/pyiree/rt/host_types.h
@@ -18,7 +18,7 @@
 #include <array>
 
 #include "absl/types/span.h"
-#include "iree/base/signature_mangle.h"
+#include "iree/base/signature_parser.h"
 #include "pyiree/rt/binding.h"
 #include "pyiree/rt/hal.h"
