+++ iree/base/BUILD
@@ -116,13 +116,11 @@ cc_library(
 )
 
 cc_library(
-    name = "signature_mangle",
-    srcs = ["signature_mangle.cc"],
-    hdrs = ["signature_mangle.h"],
+    name = "signature_parser",
+    srcs = ["signature_parser.cc"],
+    hdrs = ["signature_parser.h"],
     deps = [
-        ":status",
         "@com_google_absl//absl/container:inlined_vector",
-        "@com_google_absl//absl/memory",
         "@com_google_absl//absl/strings",
         "@com_google_absl//absl/types:optional",
         "@com_google_absl//absl/types:span",
@@ -130,10 +128,10 @@ cc_library(
 )
 
 cc_test(
-    name = "signature_mangle_test",
-    srcs = ["signature_mangle_test.cc"],
+    name = "signature_parser_test",
+    srcs = ["signature_parser_test.cc"],
     deps = [
-        ":signature_mangle",
+        ":signature_parser",
         "//iree/testing:gtest",
         "//iree/testing:gtest_main",
     ],
