+++ iree/tools/utils/CMakeLists.txt
@@ -21,7 +21,7 @@ iree_cc_library(
     absl::span
     absl::strings
     iree::base::internal::file_io
-    iree::base::signature_mangle
+    iree::base::signature_parser
     iree::base::status
     iree::hal::api
     iree::modules::hal
