+++ iree/tools/utils/vm_util.cc
@@ -22,7 +22,7 @@
 #include "absl/strings/strip.h"
 #include "absl/types/span.h"
 #include "iree/base/internal/file_io.h"
-#include "iree/base/signature_mangle.h"
+#include "iree/base/signature_parser.h"
 #include "iree/base/status.h"
 #include "iree/hal/api.h"
 #include "iree/modules/hal/hal_module.h"
